+++ build.rs
@@ -27,8 +27,9 @@ fn main() {
 
     if target_os == "android" {
         let sources = [
-            src_dir.join("BroadcastRec.java"),
             src_dir.join("InvocHdl.java"),
+            src_dir.join("BroadcastRec.java"),
+            src_dir.join("PermActivity.java"),
         ];
         let android_jar = android_build::android_jar(None);