+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes
 
+## 0.3.2
+* Added `PermissionRequest` for Android.
+* Fixed `JniClassLoader::load_class` to use `ClassLoader.loadClass` instead of `ClassLoader.findClass`.
+* Added `JniClassLoader::replace_app_loader`.
+
 ## 0.3.1
 * `jni_with_env` now clears the exception if the closure's result is `Err(Error::JavaException)`; this means any `JNIEnv` method call followed by an `?` without `.map_err(jni_clear_ex)` becomes "acceptable" (?) within `jni_with_env` closures. However, temporary structs created in the closure with `Drop` implementation doing JNI operations not listed in <https://docs.oracle.com/javase/8/docs/technotes/guides/jni/spec/design.html#exception_handling> will still *crash* the program.
 * Added `android-build` build dependency because `Dexer` is added in that crate.